Move a class to upper level.
