+++ registration/src/test/java/org/spine3/samples/lobby/registration/seat/assignment/TestSeatAssignmentsAggregate.java
@@ -1,0 +1,85 @@
+/*
+ * Copyright 2015, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package org.spine3.samples.lobby.registration.seat.assignment;
+
+import com.google.common.annotations.VisibleForTesting;
+import com.google.protobuf.Message;
+import org.spine3.samples.lobby.registration.contracts.SeatAssigned;
+import org.spine3.samples.lobby.registration.contracts.SeatAssignmentUpdated;
+import org.spine3.samples.lobby.registration.contracts.SeatAssignmentsCreated;
+import org.spine3.samples.lobby.registration.contracts.SeatAssignmentsId;
+import org.spine3.samples.lobby.registration.contracts.SeatUnassigned;
+
+import javax.annotation.Nonnull;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+
+import static com.google.common.base.Throwables.propagate;
+
+/**
+ * @author Alexander Litus
+ */
+/* package */  class TestSeatAssignmentsAggregate extends SeatAssignmentsAggregate {
+
+    /* package */ TestSeatAssignmentsAggregate(SeatAssignmentsId id) {
+        super(id);
+    }
+
+    // Is overridden to do not throw exceptions while retrieving the default state via reflection.
+    @Override
+    @SuppressWarnings("RefusedBequest")
+    protected SeatAssignments getDefaultState() {
+        return SeatAssignments.getDefaultInstance();
+    }
+
+    @VisibleForTesting
+    @Override
+    public void incrementState(@Nonnull SeatAssignments newState) {
+        super.incrementState(newState);
+    }
+
+    /* package */ void apply(SeatAssignmentsCreated event) {
+        invokeApplyMethod(event);
+    }
+
+    /* package */ void apply(SeatAssigned event) {
+        invokeApplyMethod(event);
+    }
+
+    /* package */ void apply(SeatUnassigned event) {
+        invokeApplyMethod(event);
+    }
+
+    /* package */ void apply(SeatAssignmentUpdated event) {
+        invokeApplyMethod(event);
+    }
+
+    private void invokeApplyMethod(Message event) {
+        try {
+            //noinspection DuplicateStringLiteralInspection
+            final Method apply = SeatAssignmentsAggregate.class.getDeclaredMethod("apply", event.getClass());
+            apply.setAccessible(true);
+            apply.invoke(this, event);
+        } catch (NoSuchMethodException | InvocationTargetException | IllegalAccessException e) {
+            propagate(e);
+        }
+    }
+}
